New feacture: Import same csv file again;
When a user selects the same file, the event "change" is not fired. A new attribute has added in directive: lockImportSameFile. If this is true, the directive works like today, if attr is "undefined" or "false", this directive removes the value from input [type = "file"] and unlock the "change" event.